src/requirements.txt
pymongo==4.6.0
flask-socketio==5.2.0
py-llm-core==2.5.0
flask==2.1.3
Werkzeug==2.2.2

llama_index==0.9.43
transformers
torch
pyvis==0.3.2

rake-nltk == 1.0.6

langchain==0.0.352
openai==1.6.1
arxiv==2.1.0
pypdf==3.17.4
fold-to-ascii==1.0.2.post1
pyjwt==2.8.0

keybert==0.8.3

flask_cors
python-dotenv
pydot
